requirements/pytest.txt
mock >= 3.0.0
# PyTest
pytest >= 5.4.2; sys_platform != "win32"
# https://github.com/pytest-dev/pytest/issues/6964#issuecomment-632727439
pytest >=4.6.9,<4.7; sys_platform == "win32"
pytest-salt-factories >= 0.50.4
pytest-tempdir >= 2019.10.12
pytest-helpers-namespace >= 2019.1.8
